Fix issue with incorrect text buffers
The tests were inserting text into empty buffers.
The issue was that when copying the file, the fs.copy was async.
However, that does not return a promise, thus `await` is essentially
void. Therefore, changing it to `fs.copySync` is the proper way.

Also, to prevent incorrect values for the path changes in a package,
I reordered the loading logic of the package in the before.
The package should first be activated and then the paths should be loaded.